% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\docType{data}
\name{bbc_articles}
\alias{bbc_articles}
\title{BBC articles in raw text on online gaming}
\format{A data frame with 44 bbc articles on gaming}
\usage{
data(bbc_articles)
}
\description{
A dataset containing the BBC articles text in raw text.
The variables are as follows:
}
\details{
\itemize{
  \item text bbc article in raw text
}
}
\keyword{datasets}
