% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text_features.R
\name{get_text_features}
\alias{get_text_features}
\title{Get verbatim text feature}
\usage{
get_text_features(df, text_col)
}
\arguments{
\item{df}{a dataframe/tibble}

\item{text_col}{factor/character variable that contains verbatim}
}
\description{
This is convenient wrapper function around \code{textfeatures} methods. It creates summary of text features such as how many hashtags, URLs, periods etc. are present in a verbatim row and appends and return the dataframe with textfeatures.
}
\examples{
\dontrun{
df <- data.frame(
id = c(1, 2, 3),
text = c("this is a feedback!\\t I am fedup!! https://twitter about #service @yourCare",
                 "doh", "The following list:\\n- one\\n- two\\n- three\\nOkay!?!"))
get_text_features(df,text_col = text)
}

}
\seealso{
\code{\link[textfeatures]{textfeatures}}
}
