% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_summary_cloud.R
\name{text_summary_cloud}
\alias{text_summary_cloud}
\title{Create wordcloud for a given text verbatim collection}
\usage{
text_summary_cloud(df, text_col, max_words_to_plot = 100,
  min_freq = 300)
}
\arguments{
\item{df}{a dataframe/tribble.}

\item{text_col}{text column within df}

\item{max_words_to_plot}{Maximum number of words to be plotted. least frequent terms dropped}

\item{min_freq}{words with frequency below min_freq will not be plotted}
}
\description{
A convenient wrapper function around wordcloud within wordcloud package. Breaks verbatim into word counts and plots according to user specification
}
\examples{
\dontrun{
data("text_data")
text_summary_cloud(verbatim,text_col=text,max_words_to_plot = 300,min_freq = 100)
}

}
\seealso{
\code{\link[wordcloud]{wordcloud}}
#' @seealso \code{\link[textSummary]{text_summary_cloud_by_category}}
}
