% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_summary_cloud_by_category.R
\name{text_summary_cloud_by_category}
\alias{text_summary_cloud_by_category}
\title{Create wordcloud for a given text verbatim collection, segmented by each level within a categorical variable}
\usage{
text_summary_cloud_by_category(df, text_col, categories_col,
  max_words_to_plot = 100, min_freq = 50)
}
\arguments{
\item{df}{a dataframe/tribble.}

\item{text_col}{text column within df}

\item{categories_col}{text column within df}

\item{max_words_to_plot}{Maximum number of words to be plotted. least frequent terms dropped}

\item{min_freq}{words with frequency below min_freq in the respective category_col level will not be plotted}
}
\description{
A convenient wrapper function around wordcloud method within wordcloud package. Breaks verbatim into word counts and plots according to user specification
}
\examples{
\dontrun{
data("text_data")
text_summary_cloud_by_category(verbatim,text,NPS_RATING)
}

}
\seealso{
\code{\link[wordcloud]{wordcloud}}
#' @seealso \code{\link[textSummary]{text_summary_cloud}}
}
