% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_BT.R
\name{theme_BT}
\alias{theme_BT}
\title{applies British Telecoms font and colour schemes to ggplot2 plots}
\usage{
theme_BT(axis.text.size = 12, legend.title.text.size = 17,
  axis.title.size = 14, plot.title.text.size = 20,
  plot.legend.position = "top")
}
\arguments{
\item{axis.text.size}{axis text size for both x and y axis}

\item{legend.title.text.size}{plot's legend text size}

\item{axis.title.size}{axis title text size}

\item{plot.title.text.size}{plot title text size}

\item{plot.legend.position}{plot legend position i.e. top/bottom etc.}
}
\description{
This colour scheme and font make the graphs specific to British Telecoms brand
}
\examples{
\dontrun{ggplot2() +
theme_BT()}

}
\seealso{
\code{\link[ggplot2]{theme}}

\code{\link[ggplot2]{labs}}
}
