% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model_LDA.R
\name{topic_model_LDA}
\alias{topic_model_LDA}
\title{Create topic models using LDA}
\usage{
topic_model_LDA(input_text, plot = T, number_of_topics = 4,
  number_of_top_terms_by_themes = 10)
}
\arguments{
\item{input_text}{should be a text columm from a dataframe}

\item{plot}{return a plot? TRUE by defult}

\item{number_of_topics}{number of topics/themes to return (4 by default)}

\item{number_of_top_terms_by_themes}{number of words per theme}
}
\description{
Divide collection of text into their natural groups. Get the most informative terms by a specificed number of topics, using LDA method. Please process your text column first i.e. remove stop words etc. prior to calling this function.
}
\examples{
\dontrun{
data("text_data")
top_terms_by_topic_LDA(verbatim$text,number_of_topics = 5)
verbatim \%>\% filter(NPS_RATING \%in\% c("Detractor")) \%>\%
select(text) \%>\% top_terms_by_topic_LDA(.,number_of_topics = 5)
}
}
\seealso{
\code{\link[topicmodels]{LDA}}
}
