% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_data.R
\docType{data}
\name{verbatim}
\alias{verbatim}
\title{Customer verbatims in unstructured text}
\format{A data frame with 2385 rows and 6 variables}
\usage{
data(text_data)
}
\description{
A dataset containing the feedback text and other customer related attributes such as NPS Rating of almost 2300
 customers. The variables are as follows:
}
\details{
\itemize{
  \item CUSTOMER_RESPONSE_DATE. The date when the customer submitted text
  \item SURVEY_SUB_TYPE. The journey stage where the customer is at the time of submitting feedback
  \item NPS_RATING. The customer type in terms of Net promoter Score (Promoter, Passive, Detractor)
  \item NPS_SCORE. NPS Score by the customer (0-worst) to 10 (best)
  \item text. Actual verbatim that customer submitted
}
}
\keyword{datasets}
