% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_frequencies.R
\name{word_frequencies}
\alias{word_frequencies}
\title{Return top n word frequencies in a collection of text documents}
\usage{
word_frequencies(df, text_col, number_of_words = 1, plot = TRUE,
  number_of_words_to_plot = 10, clean_text = FALSE,
  get_all_word_freqs = FALSE)
}
\arguments{
\item{df}{a tidy dataframe/tribble.}

\item{text_col}{the name of the text column within df}

\item{number_of_words}{return a plot/df of single, bigram or trigrams? returns single words by default}

\item{plot}{return a ggplot2? TRUE by default}

\item{number_of_words_to_plot}{how many words/terms to plot? Plots Top 10 words in a collection of text documents by default.}

\item{clean_text}{pre-process text? FALSE by default due to performance issues. Lammatizes and singularises words before counting}

\item{get_all_word_freqs}{return all words with their respective frequencies i.e. occurance within the whole text document i.e. whole text_col}
}
\description{
Splits the text verbatim into words, count the number of words. Either plots or returns the top n words in a collection of text documents.
}
\examples{
\dontrun{}


}
\seealso{
\code{\link[textSummary]{word_frequencies_by_category}}
}
\author{
Fahad Usman
}
